Move getSizeEstimates to Cassandra session
To reuse this method during the initialization phase in tests
